+++ Sources/SuperwallKit/Config/Models/FeatureFlags.swift
@@ -21,6 +21,7 @@ struct FeatureFlags: Decodable {
   var enableSuppressesIncrementalRendering: Bool
   var enableThrottleSchedulingPolicy: Bool
   var enableNoneSchedulingPolicy: Bool
+  var enableMultiplePaywallUrls: Bool
   var enableConfigRefresh: Bool
 
   enum CodingKeys: String, CodingKey {
@@ -42,6 +43,7 @@ struct FeatureFlags: Decodable {
     )
     enableThrottleSchedulingPolicy = rawFeatureFlags.value(forKey: "enable_throttle_scheduling_policy", default: false)
     enableNoneSchedulingPolicy = rawFeatureFlags.value(forKey: "enable_none_scheduling_policy", default: false)
+    enableMultiplePaywallUrls = rawFeatureFlags.value(forKey: "enable_multiple_paywall_urls", default: false)
     enableConfigRefresh = rawFeatureFlags.value(forKey: "enable_config_refresh", default: false)
   }
 
@@ -54,6 +56,7 @@ struct FeatureFlags: Decodable {
     enableSuppressesIncrementalRendering: Bool,
     enableThrottleSchedulingPolicy: Bool,
     enableNoneSchedulingPolicy: Bool,
+    enableMultiplePaywallUrls: Bool,
     enableConfigRefresh: Bool
   ) {
     self.enableSessionEvents = enableSessionEvents
@@ -64,6 +67,7 @@ struct FeatureFlags: Decodable {
     self.enableSuppressesIncrementalRendering = enableSuppressesIncrementalRendering
     self.enableThrottleSchedulingPolicy = enableThrottleSchedulingPolicy
     self.enableNoneSchedulingPolicy = enableNoneSchedulingPolicy
+    self.enableMultiplePaywallUrls = enableMultiplePaywallUrls
     self.enableConfigRefresh = enableConfigRefresh
   }
 }
@@ -90,7 +94,8 @@ extension FeatureFlags: Stubbable {
       disableVerboseEvents: true,
       enableSuppressesIncrementalRendering: true,
       enableThrottleSchedulingPolicy: true,
-      enableNoneSchedulingPolicy: false,
+      enableNoneSchedulingPolicy: true,
+      enableMultiplePaywallUrls: true,
       enableConfigRefresh: true
     )
   }