+++ Sources/SuperwallKit/Dependencies/DependencyContainer.swift
@@ -215,8 +215,8 @@ extension DependencyContainer: ViewControllerFactory {
     )
     let webView = SWWebView(
       isMac: deviceHelper.isMac,
-      sessionEventsManager: sessionEventsManager,
       messageHandler: messageHandler,
+      isOnDeviceCacheEnabled: paywall.onDeviceCache == .enabled,
       factory: self
     )
     let paywallViewController = PaywallViewController(