+++ Sources/SuperwallKit/Models/Paywall/Paywall.swift
@@ -4,7 +4,7 @@
 //
 //  Created by Yusuf Tör on 28/02/2022.
 //
-// swiftlint:disable function_body_length type_body_length
+// swiftlint:disable function_body_length type_body_length file_length
 
 import UIKit
 
@@ -24,6 +24,9 @@ struct Paywall: Decodable {
 
   /// The URL of the paywall webpage
   var url: URL
+
+  /// An array of potential URLs to load the paywall from.
+  let urlConfig: WebViewURLConfig
 
   /// Contains the website modifications that are made on the paywall editor to be accepted
   /// by the webview.
@@ -114,6 +117,7 @@ struct Paywall: Decodable {
     case identifier
     case name
     case url
+    case urlConfig
     case htmlSubstitutions = "paywalljsEvent"
     case presentationStyle = "presentationStyleV2"
     case presentationCondition
@@ -147,6 +151,7 @@ struct Paywall: Decodable {
     identifier = try values.decode(String.self, forKey: .identifier)
     name = try values.decode(String.self, forKey: .name)
     url = try values.decode(URL.self, forKey: .url)
+    urlConfig = try values.decode(WebViewURLConfig.self, forKey: .urlConfig)
     htmlSubstitutions = try values.decode(String.self, forKey: .htmlSubstitutions)
 
     let throwableSurveys = try values.decodeIfPresent(
@@ -261,6 +266,7 @@ struct Paywall: Decodable {
     name: String,
     experiment: Experiment? = nil,
     url: URL,
+    urlConfig: WebViewURLConfig,
     htmlSubstitutions: String,
     presentation: PaywallPresentationInfo,
     backgroundColorHex: String,
@@ -288,6 +294,7 @@ struct Paywall: Decodable {
     self.name = name
     self.experiment = experiment
     self.url = url
+    self.urlConfig = urlConfig
     self.htmlSubstitutions = htmlSubstitutions
     self.presentation = presentation
     self.backgroundColor = backgroundColor
@@ -372,6 +379,7 @@ extension Paywall: Stubbable {
       identifier: "identifier",
       name: "abc",
       url: URL(string: "https://google.com")!,
+      urlConfig: .init(endpoints: [], maxAttempts: 0),
       htmlSubstitutions: "",
       presentation: PaywallPresentationInfo(
         style: .modal,